🐛 [RUMF-1296] ignore non-primary pointer events
In 'listenActionEvents', we expect that a 'click' event is preceded by
a single 'pointerdown' event. It turns out this is not always the case
on multitouch devices.

This issue would be amplified when listening to 'pointerup' events.

Let's just focus on the 'primary' pointer events. This will ignore
secondary touches until we implement proper support for them (if ever).